refactor: 프로필 페이지 리팩토링 (드롭다운 + 팔로우 컴포넌트 분리)
