+++ src/constants/Profile.ts
@@ -1,0 +1,4 @@
+export const ALTS = {
+  UN_FOLLOW: '구독 취소하기',
+  FOLLOW: '구독하기',
+} as const;
